Refactor controller package
