+++ src/main/java/tecna/controller/DialogBox.java
@@ -1,4 +1,4 @@
-package tecna;
+package tecna.controller;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -29,7 +29,7 @@ public class DialogBox extends HBox {
 
     private DialogBox(String text, Image img) {
         try {
-            FXMLLoader fxmlLoader = new FXMLLoader(tecna.MainWindow.class.getResource("/view/DialogBox.fxml"));
+            FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
             fxmlLoader.setController(this);
             fxmlLoader.setRoot(this);
             fxmlLoader.load();
